added simple workaround around issue with username
Currently, the 'postAuthor' field inadvertedly includes part of the caption if there is on one. As a simple workaround, this checks if the postAuthor field has a space in it ( usernames aren't allowed to have spaces in them ), and only if it doesnt will the username will be included in the resulting filename for the image saved.